requirements.txt
accelerate
torchvision
transformers>=4.25.1
ftfy
tensorboard
modelcards
datasets
# diffusers==0.11.1
-e ./diffusers